Remove unnecessary unboxing

+++ src/components/src/main/java/org/apache/jmeter/extractor/RegexExtractor.java
@@ -295,7 +295,7 @@ public class RegexExtractor extends AbstractScopedTestElement implements PostPro
                 log.debug("RegexExtractor: Template piece {} ({})", obj, obj.getClass());
             }
             if (obj instanceof Integer) {
-                result.append(match.group(((Integer) obj).intValue()));
+                result.append(match.group((Integer) obj));
             } else {
                 result.append(obj);
             }
