+++ src/core/src/main/java/org/apache/jmeter/control/LoopController.java
@@ -81,10 +81,10 @@ public class LoopController extends GenericController implements Serializable, I
         // Evaluation occurs when nbLoops is not yet evaluated
         // or when nbLoops is equal to special value INFINITE_LOOP_COUNT
         if (nbLoops==null || // No evaluated yet
-                nbLoops.intValue()==0 || // Last iteration led to nbLoops == 0,
+                nbLoops ==0 || // Last iteration led to nbLoops == 0,
                                          // in this case as resetLoopCount will not be called,
                                          // it leads to no further evaluations if we don't evaluate, see BUG 56276
-                nbLoops.intValue()==INFINITE_LOOP_COUNT // Number of iteration is set to infinite
+                nbLoops ==INFINITE_LOOP_COUNT // Number of iteration is set to infinite
                 ) {
             try {
                 JMeterProperty prop = getProperty(LOOPS);
@@ -93,7 +93,7 @@ public class LoopController extends GenericController implements Serializable, I
                 nbLoops = 0;
             }
         }
-        return nbLoops.intValue();
+        return nbLoops;
     }
 
     public String getLoopString() {
