+++ src/jorphan/src/main/java/org/apache/jorphan/gui/MinMaxLongRenderer.java
@@ -38,7 +38,7 @@ public class MinMaxLongRenderer extends NumberRenderer { // NOSONAR
     @Override
     public void setValue(Object value) {
         if (value instanceof Long) {
-            long longValue = ((Long) value).longValue();
+            long longValue = (Long) value;
             if (!(longValue == Long.MAX_VALUE || longValue == Long.MIN_VALUE)) {
                 setText(formatter.format(longValue));
                 return;
