+++ src/protocol/http/src/main/java/org/apache/jmeter/protocol/http/control/HeaderManager.java
@@ -216,7 +216,7 @@ public class HeaderManager extends ConfigTestElement implements Serializable, Re
             }
         }
         for (Integer indice : removeIndices) {
-            getHeaders().remove(indice.intValue());
+            getHeaders().remove(indice);
         }
     }
